Made it more navigatable
